Refactor Bubble environments initialization
