+++ GliaWidgets/Sources/Component/Bubble/BubbleWindow.swift
@@ -123,12 +123,7 @@ class BubbleViewController: UIViewController {
     }
 }
 
-extension BubbleWindow {
-    struct Environment {
-        var uiScreen: UIKitBased.UIScreen
-        var uiApplication: UIKitBased.UIApplication
-    }
-}
+
 
 #if DEBUG
 extension BubbleWindow {
@@ -140,10 +135,6 @@ extension BubbleWindow {
         }
         return window
     }
-}
-
-extension BubbleWindow.Environment {
-    static let mock = Self(uiScreen: .mock, uiApplication: .mock)
 }
 
 #endif
